Make test invalid_header_block_fragment more readable
