+++ crates/httpwg/src/rfc9113/mod.rs
@@ -27,9 +27,6 @@ pub fn default_settings() -> SettingPairs<'static> {
 
 pub mod _3_4_http2_connection_preface;
 
-pub mod _4_1_frame_format;
-pub mod _4_2_frame_size;
-pub mod _4_3_header_compression_and_decompression;
-
+pub mod _4_http_frames;
 pub mod _5_streams_and_multiplexing;
 pub mod _6_frame_definitions;
